use js prompt to enter name

+++ src/front-end/lists/static/js/custom/pages/home/list-settings.js
@@ -175,10 +175,12 @@ export class ListSettings
     **********************************************************/
     async delete() {
         const listDelete = new ListDelete(this.listID);
-        listDelete.delete();
 
-        listDelete.removeListElements();
-        ListSettings.closeModal();
+        if (listDelete.confirm()) {
+            listDelete.delete();
+            listDelete.removeListElements();
+            ListSettings.closeModal();
+        }
     }
 
     /**********************************************************
